changes in structues
